# Conflicts:
#	Implementierung/studyplan_server/src/main/java/edu/kit/informatik/studyplan/server/model/moduledata/Module.java
#	Implementierung/studyplan_server/src/main/java/edu/kit/informatik/studyplan/server/rest/resources/PlansResource.java

+++ Implementierung/studyplan_server/src/main/java/edu/kit/informatik/studyplan/server/filter/condition/Condition.java
@@ -61,7 +61,7 @@ public final class Condition {
      * @param upper the upper bound
      * @return the condition
      */
-    public static Condition createBetween(String lhsName, int lower, int upper) {
+    public static Condition createBetween(String lhsName, double lower, double upper) {
         return new Condition(lhsName, new Object[]{lower, upper}, BooleanRelation.BETWEEN);
     }
